# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        items:
          $ref: '#/definitions/domain.FieldError'
        type: array
      message:
        example: Validation failed
        type: string
    type: object
  domain.FieldError:
    properties:
      field:
        example: email
        type: string
      message:
        example: This field is required.
        type: string
    type: object
...
  domain.TokensPair:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
host: petbackend-a2vg.onrender.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a PetHelp project backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /oauth/{provider}:
    get:
      consumes:
      - application/json
      description: Handles the callback from an OAuth2.0 provider after user authentication.
      parameters:
      - description: OAuth2.0 provider name (e.g., google, github)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Successfully authenticated and generated tokens
          schema:
            $ref: '#/definitions/domain.TokensPair'
        "404":
          description: Account with this email not found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error or failed to complete OAuth2.0 authentication
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: OAuth2.0 Provider Callback
      tags:
      - OAuth
...
  /token/refresh:
    post:
      description: Exchanges a valid refresh token (from an HTTP-only cookie) for
        a new access token and a new refresh token. The used refresh token is revoked.
        This endpoint does not accept a request body.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated new token pair
          schema:
            $ref: '#/definitions/domain.TokensPair'
        "400":
          description: Invalid request payload or malformed refresh token
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: 'Unauthorized: Invalid, expired, or malformed refresh token'
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "403":
          description: 'Forbidden: Refresh token has been revoked or is otherwise
            not allowed'
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update access and refresh tokens
      tags:
      - Token
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"